W225: eighty percent — +3 MATCH (6562 -> 6565, 80.0%), nd -333
First-party matching crosses **80.0%** (6565 of 8211). Third main-model template-transfer wave;
0 MATCHes lost, 0 new over-window, 0 regressions kept.

**The three matches:**

  btlAction 0028b800    nd27 -> **0**.  The function already had the sibling's staging and case
                        order; what remained was hand-computed field addressing where the sibling
                        writes `enemyRecords[unitId].moveSpeed[allowMove].speedIndex`.
  mdlFile 00326cf0      nd2 -> **0**.
  _h_camp_data 0016c010 -> **0**.

**Typed indexing has now matched in two unrelated files.** `mdlFile 00324510` (nd76 -> 0) and
`btlAction 0028b800` (nd27 -> 0) were both `*(T*)((u8*)base + i*STRIDE + OFF)` where the matched
sibling names a real array of a real struct. The compiler indexes a declared array more compactly
than we can spell by hand, so it removes bytes as well as matching, and the sibling supplies the
actual type rather than an invented one. It was broadcast mid-wave and is now in the idiom library.

Largest gain: `fclShopMisc 003fb2f0` **nd136 -> 18**, from taking four shapes off its sibling at
once — positive populated-entry branches, default-first clamp updates, unsigned money division, and
an immediate `return 0` for an unchanged quantity. Also `mdlFile 0034e4c0` nd27 -> **6**,
`itfPanel 003ade70` nd19 -> **4** (duplicated `u16` staging removed from both branches so the cast
survives only on the non-0xff path), `mdlFile 003268c0` nd54 -> 28, `op_menu 0026cc90` nd1118 -> 1089
(seven nested sprite/raster expressions replaced by the accessor its sibling calls directly),
`k_data 001b8db0` nd327 -> 309, `datCalc 0030af00` nd210 -> 198.

**This wave was mostly negatives, and they are the point.** Roughly four in five transfers were
measured and restored, which is expected at this depth and steadily narrows what remains. Several
close off lines that would otherwise be retried indefinitely:

- `mt_evtMenu 00381c60`: the draw-call live-value idiom helps case 0 by 14 and regresses cases 1
  through 7 by exactly 16 each. Seven cases tested individually. **It is per-case, not per-switch.**
- `bp_root 0022ae80`: converting to the sibling's absolute-array addressing regresses 1038 -> 1213,
  proving the GP-relative scalar spelling is already correct there.
- `bcm_panel 0022df10`, `00225040`, `00229b40`: the phase-boundary pointer rematerialization that
  paid 192 points at `00227800` regresses at all three. **Its boundaries are function-specific.**
- `k_clump 001a8140` and `001a8b10`: absolute camera and function-pointer addressing has now
  regressed three separate times. Closed.
- `fclShopMisc 003f12a0` sits at **nd1** and full fndiff confirms the sole remaining semantic word is
  `addu v0,v1,v0` against retail's `addu v0,v0,v1` — a commutative operand floor. It is as finished
  as it can be.

Six of the eight slices reported their nearest targets are now single-word commutative-operand or
cyclic-colouring residuals with matching save counts, which is the measured floor for this compiler.

+++ src/Main/Battle/Data/datCalc.c
@@ -5952,7 +5952,7 @@ u32 FUN_0030af00_narrow(u16 *param_1, u16 *param_2, u16 param_3,
                           u16 param_4, u32 param_5)
 {
   u16 uVar1;
-  short sVar2;
+  int sVar2;
   int iVar3;
   u32 uVar4;
   u16 *puVar5;
